requirements.txt
# Flask and Flask Extensions
Flask==2.3.3
Flask-SocketIO==5.3.5
Flask-Cors==4.0.0

# Computer Vision and Machine Learning
opencv-python-headless==4.8.0.74
cvzone==1.5.6
ultralytics==8.0.130
numpy==1.24.3
Pillow==10.0.1
imutils==0.5.4

# Base64 Encoding/Decoding (Python Standard Library, for reference)

# Logging (Python Standard Library, for reference)

# Asynchronous Support
eventlet==0.33.3  # Used by Flask-SocketIO for async (alternative to gevent)

# Optional: Development and Debugging Tools
gunicorn==21.1.0  # Production WSGI server (if used)
